SCADA: refactor
