+++ OutageManagementSystem/OutageManagementSystem.sln
@@ -73,7 +73,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PubSubEngine", "PubSubEngin
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SCADA", "SCADA", "{F47B4360-F4CF-424E-A85E-623A1D2F297B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModBus", "SCADA\ModBus\ModBus.csproj", "{6E88F6B5-11CB-485A-9BDF-C622FF837022}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Modbus", "SCADA\ModBus\Modbus.csproj", "{6E88F6B5-11CB-485A-9BDF-C622FF837022}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SCADAData", "SCADA\SCADA_Config_Data\SCADAData.csproj", "{C37F054D-37D9-46C0-881C-1EE3C8C77633}"
 EndProject
@@ -115,6 +115,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OMSCommon", "OutageManagementService\OMSCommon\OMSCommon.csproj", "{BC054AA3-2DD3-4657-B27B-6C2126803106}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common.Cloud", "..\OMS.Cloud\Common\Common.Cloud\Common.Cloud.csproj", "{FBEE1369-0F6C-41E6-8C62-8C1C5CDBB602}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ModbusFunctions", "SCADA\ModbusFunction\ModbusFunctions.csproj", "{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -530,6 +532,18 @@ Global
 		{FBEE1369-0F6C-41E6-8C62-8C1C5CDBB602}.Release|x64.Build.0 = Release|Any CPU
 		{FBEE1369-0F6C-41E6-8C62-8C1C5CDBB602}.Release|x86.ActiveCfg = Release|Any CPU
 		{FBEE1369-0F6C-41E6-8C62-8C1C5CDBB602}.Release|x86.Build.0 = Release|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Debug|x64.Build.0 = Debug|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Debug|x86.Build.0 = Debug|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Release|x64.ActiveCfg = Release|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Release|x64.Build.0 = Release|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Release|x86.ActiveCfg = Release|Any CPU
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -572,6 +586,7 @@ Global
 		{5272C603-5EFF-4248-BB3D-563D05B48185} = {9DB302A8-78A0-459E-99DD-FE5E80E93E24}
 		{BC054AA3-2DD3-4657-B27B-6C2126803106} = {9DB302A8-78A0-459E-99DD-FE5E80E93E24}
 		{FBEE1369-0F6C-41E6-8C62-8C1C5CDBB602} = {4C80DEF4-45DA-49FD-A732-0D686DC648EE}
+		{3A07C9C5-D4C1-4DB2-9E8B-755759FED39C} = {F47B4360-F4CF-424E-A85E-623A1D2F297B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {28760A05-312F-4AA5-9831-305EE698F2F1}
