+++ OutageManagementSystem/SCADA/ModbusFunction/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EasyModbusTCP" version="5.5.0" targetFramework="net48" />
+</packages>
